move photon/cache/ to photon/fs/cache/
