--- cache/ocf_cache/include/ocf
@@ -1,1 +1,0 @@
-../../../build/_deps/ocf_lib-src/inc/
