+++ fs/cache/ocf_cache/CMakeLists.txt
@@ -17,7 +17,7 @@ file(GLOB_RECURSE src_ocf_cache ocf_cache.cpp ocf_namespace.cpp
   photon_bindings/*.cpp
 )
 add_library(ocf_cache_lib STATIC ${src_ocf_cache})
-target_include_directories(ocf_cache_lib PUBLIC include/ photon_bindings/env/ ../../include/)
+target_include_directories(ocf_cache_lib PUBLIC include/ photon_bindings/env/ ../../../include/)
 target_link_libraries(ocf_cache_lib PRIVATE ocf_lib)
 
 # if(PHOTON_BUILD_TESTING)
