+++ fs/cache/store.cpp
@@ -14,8 +14,8 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-#include <photon/cache/pool_store.h>
-#include <photon/cache/cache.h>
+#include <photon/fs/cache/pool_store.h>
+#include <photon/fs/cache/cache.h>
 #include <photon/common/alog.h>
 #include <photon/common/alog-stdstring.h>
 #include <photon/common/alog-audit.h>
