--- include/photon/cache/cache.h
@@ -1,1 +1,0 @@
-../../../cache/cache.h
