--- include/photon/cache/forwardcfs.h
@@ -1,1 +1,0 @@
-../../../cache/forwardcfs.h
