--- include/photon/cache/pool_store.h
@@ -1,1 +1,0 @@
-../../../cache/pool_store.h
