+++ include/photon/fs/cache/cache.h
@@ -1,0 +1,1 @@
+../../../../fs/cache/cache.h
